According to the compiled alarm repair code

+++ services/media_engine/plugins/ffmpeg_adapter/common/ffmpeg_converter.cpp
@@ -428,7 +428,8 @@ std::string FFMpegConverter::AVStrError(int errnum)
 AudioChannelLayout FFMpegConverter::ConvertAudioVividToOHAudioChannelLayout(uint64_t ffChannelLayout, int channels)
 {
     auto ite = std::find_if(g_audioVividChannelLayoutMap.begin(), g_audioVividChannelLayoutMap.end(),
-                            [&ffChannelLayout](const auto &item) -> bool { return item.first == ffChannelLayout;});
+                            [&ffChannelLayout](const auto &item) -> bool {
+                                return static_cast<uint64_t>(item.first) == ffChannelLayout;});
     if (ite == g_audioVividChannelLayoutMap.end() || ite -> second != channels) {
         MEDIA_LOG_W("Convert channel layout failed: " PUBLIC_LOG_U64, ffChannelLayout);
         return GetDefaultChannelLayout(channels);
